pipeline/pyproject.toml
[project]
name = "muni-walk-access"
version = "0.1.0"
description = "SF MUNI walkshed accessibility pipeline"
requires-python = ">=3.12"
dependencies = [
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "duckdb>=1.0",
    "polars>=1.0",
    "httpx>=0.28.1",
]

# Geo deps require: brew install gdal geos proj  (added in Story 1.5)
[project.optional-dependencies]
geospatial = [
    "osmnx>=2.0",
    "pandana>=0.7",
    "geopandas>=1.0",
    "pyproj>=3.7",
    "shapely>=2.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "ruff>=0.8",
    "mypy>=1.14",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/muni_walk_access"]

[tool.ruff]
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "I", "N", "D"]
ignore = ["D100", "D104"]

[tool.mypy]
strict = true

[[tool.mypy.overrides]]
module = "yaml"
ignore_missing_imports = true
